requirements.txt
streamlit
langchain
langchain_community
sentence-transformers
huggingface_hub
PyMuPDF
python-dotenv
faiss-cpu

